Add redirect to proposals list page - use LayoutBalanceSize() for treasury balance display - add proposal date to 'in discussion' proposal on layoutAuthorAndDate - Add redirect to proposals details page from overview page - Add dividing lines between proposals - Clean up IconEditor widget - Add search editor, - Add proposal category and orderby dropdown - restore sync icon with updated dimensions - clean up code
