+++ ui/page/transaction/transactions_page.go
@@ -172,7 +172,7 @@ func (pg *TransactionsPage) Layout(gtx layout.Context) layout.Dimensions {
 				return pg.orderDropDown.Layout(gtx, 0, true)
 			}),
 			layout.Expanded(func(gtx C) D {
-				return pg.txTypeDropDown.Layout(gtx, pg.orderDropDown.Width, true)
+				return pg.txTypeDropDown.Layout(gtx, pg.orderDropDown.Width-6, true)
 			}),
 		)
 	}
